Fixed some tag mistakes and added style to the texts as well as background.
I fixed a typo with the div tag and changed the logo image. In addition, font type and color was added to the different texts. I also choose different parameters such as the width, sizes and positions for the texts. Finally, I added a background image and used flex, border-box and other properties to style the image.