Added the section "Our Story" and gave some style using CSS
The things that were changed:
background color, certain font colors

The things that were added:
Division classes were added in order to organize the website. Specifically, headline and sub-headlines were given styling properties such as letter-spacing, text-transform, font-size and color. In addition, a margin was put in place for the different texts and the buttons. (cta-btn, body-btn).

The section called "Our Story" was created in order to place some information about the restaurant later on.
Also, the property" hover "was used in order to change the color of the buttons when clicked.