Added 2 new sections and styled them.
I created a section called "Discover Our Menu" and grouped the 4 images above the section, using grid-template. I gave some dimensions like padding, margins and also included a button that lead to the menu page when clicked.

The second section is called the "Caldera-View" and I added a background image of the Caldera volcano in Greece because it can be seen while looking out from the restaurant.