Created menu page
I added different pictures that are part of the menu and gave them short captions and used grid style to make columns.  I created three different sections altogether. The first one is the introduction section to the meals at Ambrosia. The second section is a 4 image gallery and the third section is a grid style collection of menu items.